% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeSCED.R
\name{fit_and_transform_model}
\alias{fit_and_transform_model}
\title{Fit and Transform Model}
\usage{
fit_and_transform_model(
  data,
  formula,
  random_formula,
  ar1_formula,
  time_col,
  phase_col,
  interaction_col,
  id_col
)
}
\arguments{
\item{data}{A data frame containing the data to be modeled.}

\item{formula}{A formula for the fixed effects in the initial model.}

\item{random_formula}{A formula for the random effects in the initial model.}

\item{ar1_formula}{A formula for the AR(1) correlation structure in the initial model.}

\item{time_col}{The name of the time variable.}

\item{phase_col}{The name of the phase variable.}

\item{interaction_col}{The name of the interaction variable.}

\item{id_col}{The name of the ID variable.}

\item{nmeasurement_list}{The number of measurements per subject.}

\item{nsubject_list}{The number of subjects in the dataset.}
}
\value{
A fitted mixed-effects model from the \code{lmerTest} package.
}
\description{
This function fits an initial linear mixed-effects model using the \code{nlme} package, applies a transformation based on the fitted model, and then fits a transformed model using the \code{lmerTest} package.
}
\examples{
\dontrun{
fit_and_transform_model(data, y ~ time + phase + interaction, ~ 1 + time + phase + interaction | id, ~ time | id, "time", "phase", "interaction", "id")
}
}
