% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_sced.R
\name{lme_sced}
\alias{lme_sced}
\title{Fit a Linear Mixed Effects Model for SCED Data}
\usage{
lme_sced(data, fixed, random, control = NULL, ar_1 = NULL)
}
\arguments{
\item{data}{A data frame containing the variables for the model.}

\item{fixed}{A formula specifying the fixed effects.}

\item{random}{A formula specifying the random effects.}

\item{control}{A list of control options for the model (optional).}

\item{ar_1}{A formula specifying the AR(1) correlation structure (optional).}
}
\value{
A model object or a list containing the model and AR(1) parameter value.
}
\description{
This function fits a linear mixed effects model for SCED data, with the option to account for AR(1) correlation.
}
\examples{
# Example usage:
# model <- lme_sced(data, fixed = y ~ time + phase, random = ~ phase | id, ar_1 = ~ time | id)


}
