% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlrt.R
\name{rlrt}
\alias{rlrt}
\title{Restricted Likelihood Ratio Test for Variance Components}
\usage{
rlrt(lme_sced_results)
}
\arguments{
\item{lme_sced_results}{A list object returned by the \code{lme_sced} function.}
}
\value{
A list containing:
\item{full_model_nlme}{The full model fitted with nlme.}
\item{diag_model_nlme}{The reduced model (diagonal covariance) fitted with nlme.}
\item{p_values_random_effects}{P-values for the variance components (BC-RLRT).}
\item{p_value_covariance}{P-value for the covariance components (Parametric Bootstrap).}
}
\description{
Performs a boundary-corrected restricted likelihood ratio test (BC-RLRT) for variance components
and a parametric bootstrap for covariance components in SCED models.
}
