% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_sced.R
\name{lme_sced}
\alias{lme_sced}
\title{Fit and Transform Model}
\usage{
lme_sced(
  data,
  formula,
  random_formula,
  ar1_formula,
  time_col,
  phase_col,
  interaction_col,
  id_col
)
}
\arguments{
\item{data}{Data frame containing the dataset.}

\item{formula}{A formula for the fixed effects in the initial model.}

\item{random_formula}{A formula for the random effects in the initial model.}

\item{ar1_formula}{A formula for the AR1 correlation structure in the initial model.}

\item{time_col}{The name of the time column in the dataset.}

\item{phase_col}{The name of the phase column in the dataset.}

\item{interaction_col}{The name of the interaction column in the dataset.}

\item{id_col}{The name of the ID column in the dataset.}
}
\value{
The fitted lmer model after transformations.
}
\description{
This function fits an initial model using nlme::lme, performs transformations, and then fits a second model using lmerTest::lmer.
}
\examples{
# Assuming `data` is a data frame with appropriate columns
# lme_sced(data, y ~ time + phase + interaction, ~ 1 + time + phase + interaction | id, ~ time | id, "time", "phase", "interaction", "id")


}
